R08/T02/feature/register_restaurant_page2 | Refatoração de código
- Criei uma classe para o map
- Mudei nomes de arquivos
